backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.1
python-dotenv==1.0.0
numpy==1.26.4
